% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection_list_by_project}
\alias{collection_list_by_project}
\title{List all collections for a project}
\usage{
collection_list_by_project(id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{id}{\strong{(required)} Numeric project id}
}
\value{
list
}
\description{
List all collections in which a project is included
}
