% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection_remove_projects}
\alias{collection_remove_projects}
\title{Remove projects from a collection}
\usage{
collection_remove_projects(
  collection_id,
  projects = c(),
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{collection_id}{(required) Collection ID}

\item{projects}{\strong{(required)} List of projects}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Remove projects from a collection
}
\examples{


collection_remove_projects (
  collection_id=123,
  projects = c("unique-idno-for-project", "another-project-idno")
)




}
