% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create new project}
\usage{
create_project(
  type,
  idno,
  metadata,
  thumbnail = NULL,
  collection_ids = list(),
  collection_names = list(),
  overwrite = FALSE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{type}{(required) Type of project - survey, geospatial, table, document, timeseries}

\item{idno}{\strong{(required)} Project unique IDNO}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}

\item{thumbnail}{Path to thumbnail image file. Supported types are JPG, JPEG, GIF, PNG}

\item{collection_names}{list of collection Names to assign to project}

\item{overwrite}{Overwrite an existing project? TRUE | FALSE}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}

\item{collection_id}{List of collection IDs to assign to project}
}
\description{
Create a new project
}
\examples{


create (
  type="survey",
  idno = "unique-idno-for-project",
  metadata = list()
)




}
