% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{import_project}
\alias{import_project}
\title{Import project}
\usage{
import_project(type, file_path, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{type}{(required) Type of project - survey, geospatial, table, document, timeseries}

\item{file_path}{Path to project file. Supported types are DDI/XML, JSON or project package zip}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Import project
}
\examples{


import_project (
  type="timeseries",
  file_path = "/files/some-project.zip"
)




}
