% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{list_collections}
\alias{list_collections}
\title{Get collections list}
\usage{
list_collections(api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\value{
list
}
\description{
Get collections list
}
\examples{


list_collections (
)




}
