% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{list_projects}
\alias{list_projects}
\title{Get paginated projects list}
\usage{
list_projects(
  keywords = NULL,
  filter_type = c(),
  filter_collection = c(),
  limit = 100,
  offset = 0,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{keywords}{Search keywords}

\item{filter_type}{Filter by project type e.g. c("timeseries")}

\item{filter_collection}{Filter by collections e.g. c(1,2,3)}

\item{limit}{Limit number of search entries. Default is 100}

\item{offset}{Starting row number for pagination}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}

\item{type}{(required) Type of project - survey, geospatial, table, document, timeseries}
}
\description{
Get paginated projects list
}
\examples{


list_projects (
  keywords="some keywords or project IDNO",
  filter_type = c("survey","timeseries"),
  filter_collection = c(1,2)
  metadata = list()
)




}
