% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metadataedit_http_delete}
\alias{metadataedit_http_delete}
\title{Make an http DELETE request}
\usage{
metadataedit_http_delete(
  url,
  options = list(),
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{url}{}
}
\value{
http request output
}
\description{
Make an http DELETE request
}
