% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metadataedit_http_get}
\alias{metadataedit_http_get}
\title{Make an http GET request}
\usage{
metadataedit_http_get(
  url,
  options = list(),
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{url}{}
}
\value{
http request output
}
\description{
Make an http GET request
}
