% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_by_id}
\alias{project_by_id}
\title{Find a project by ID}
\usage{
project_by_id(id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{id}{(required) Project numeric ID}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Find a project by ID
}
\examples{

project_by_id (
  id=12
)

}
