% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resources_delete}
\alias{resources_delete}
\title{Editor - delete External Resources}
\usage{
resources_delete(idno, resource_id, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{Project IDNo}

\item{resource_id}{Resource ID}
}
\value{
list
}
\description{
Delete external resources for a project
}
