% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resources_list}
\alias{resources_list}
\title{Editor - list External Resources}
\usage{
resources_list(idno, api_key = NULL, api_base_url = NULL)
}
\arguments{
\item{idno}{Project IDNo}
}
\value{
List of external resources
}
\description{
List external resources for a project
}
