% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resources_update}
\alias{resources_update}
\title{Editor update resource}
\usage{
resources_update(
  idno,
  resource_id,
  dctype = NULL,
  title = NULL,
  subtitle = NULL,
  dcformat = NULL,
  author = NULL,
  dcdate = NULL,
  country = NULL,
  language = NULL,
  contributor = NULL,
  publisher = NULL,
  rights = NULL,
  description = NULL,
  abstract = NULL,
  toc = NULL,
  subjects = NULL,
  file_path = NULL,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{idno}{\strong{(required)} Project IDNO}

\item{resource_id}{\strong{(required)} Resource ID}

\item{dctype}{Resource document type}

\item{title}{Resource title}

\item{dcformat}{Resource file format}

\item{author}{Author name}

\item{dcdate}{Date using YYYY-MM-DD format}

\item{country}{Country name}

\item{language}{Language or Language code}

\item{contributor}{Contributor name}

\item{publisher}{Publisher name}

\item{rights}{Rights}

\item{description}{Resource detailed description}

\item{abstract}{Resource abstract}

\item{toc}{Table of contents}

\item{file_path}{File path for uploading}

\item{overwrite}{Overwrite if resource already exists? Accepted values "yes", "no"}
}
\description{
Update resource
}
