% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{update_project}
\alias{update_project}
\title{Update project}
\usage{
update_project(
  type,
  idno,
  metadata,
  partial_update = FALSE,
  thumbnail = NULL,
  overwrite = FALSE,
  api_key = NULL,
  api_base_url = NULL
)
}
\arguments{
\item{type}{(required) Type of project - survey, geospatial, table, document, timeseries}

\item{idno}{\strong{(required)} Project unique IDNO}

\item{metadata}{\strong{(required)} Metadata list depending on the type of study}

\item{partial_update}{Update only partial metadata (TRUE/FALSE)}

\item{api_key}{API key (optional if API key is set using set_api_key)}

\item{api_base_url}{API base endpoint (optional if API base endpoint is set using set_api_url)}
}
\description{
Update project
}
\examples{


update_project (
  type="survey",
  idno = "unique-idno-for-project",
  metadata = list()
)




}
